Tia/requirements.txt
openai-whisper==20231117
git+https://github.com/openai/CLIP.git
torch
torchvision
torchaudio
opencv-python
matplotlib
Pillow
numpy